% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakdete.R
\name{peak_DETE}
\alias{peak_DETE}
\title{Peak Detection}
\usage{
peak_DETE(x, m)
}
\arguments{
\item{x}{Input signal}

\item{m}{minimum peak-to-peak separation distance ( in indices)}
}
\value{
Array representing  peak points
}
\description{
Find peak and bottom points of a time series signal.
}
\details{
This function to determine where a peak or bottom points is located  in a time series, from beginning to end.
}
\examples{
t<-seq(0,1,.001)

x<-sin(2*pi*20*t)

peak_DETE(x, 30)

peak_DETE(-x, 30)

## Not run:

t<-seq(0,1,.001)
x<-sin(2*pi*20*t)
plot(x, type="l", col="navy")
peaks <- peak_DETE(x, 30)
bottom<- peak_DETE(-x, 30)
points(peaks,x[peaks],pch=19,col=2)
points(bottom,x[bottom],pch=19,col=3)

## End(Not run)

}
